refactor: 마이페이지 뷰 분리
