+++ accounts/views/__init__.py
@@ -2,3 +2,4 @@ from .auth_dev_views import *
 from .auth_user_views import *
 from .token_views import *
 from .user_info_views import *
+from .my_page_views import *
